Refactor CLI auth into provider modules
